Fix spurious pandas chained-assignment warnings in Cython frame builders
Build frames via constructor/.assign; consolidate regression tests.
